requirement.txt
numpy
pytorch = 1.13.1
torchvision = 0.14.1
pandas
tqdm
pillow
scikit-learn